Defined lanes and adjusted the cars position ..
.. when toggling between fullscreen and windowed mode.
It took some testing to find out the position of the lanes.
All future introduced elements will have to have their y position adjusted, since the fullscreen mode does not just scale, but draws the background image at a different position! Luckily this is simple since you only need to add or subtract the value "ACTUAL_SCREEN_HEIGHT // 8" to/from the y position of the element