introduced new unit for positioning elements
The unit is one percent of the actual screen width,. The second unit is one percent of the actual screen height. I went through the complete code and adjusted every element positioning to calculate with this unit - hopefully got them all - will test on different screen later.
Also I deleted some code related to windowed mode, since we are not going to use it in this way anymore.